Added insert and get stored procs, and corrosponding controller and data controller methods

+++ DotaBrackets/DotaBrackets_WEB_2016/Models/Gamer.cs
@@ -12,7 +12,7 @@ namespace DotaBrackets_WEB_2016.Models
         public int preferencesID { get; set; }
         public string userName { get; set; }
         public string access { get; set; }
-        public int steamID { get; set; }
+        public long steamID { get; set; }
         public int dota2ID { get; set; }
         public bool isSearching { get; set; }
         public string avatar { get; set; }
